+++ app/Services/Facebook/Client.php
@@ -3,13 +3,16 @@
 namespace App\Services\Facebook;
 
 use FacebookAds\Api;
+use FacebookAds\Object\IGUser;
+use FacebookAds\Object\Page;
+use FacebookAds\Object\User;
 use Illuminate\Support\Collection;
 
 class Client
 {
     protected Api $api;
 
-    public function __construct(protected string $access_token)
+    public function __construct(protected string $access_token, protected ?string $user_id = null)
     {
         $this->api = Api::init(config('services.facebook.client_id'), config('services.facebook.client_secret'), $access_token);
         $this->api->setDefaultGraphVersion('24.0');
@@ -27,29 +30,22 @@ class Client
 
     public function accounts(): Collection
     {
-        $accounts = $this->api->call('/me/accounts', params: [
-            'fields' => '
-                id,
-                name,
-                access_token,
-                category,
-                followers_count,
-                verification_status,
-                instagram_business_account{
-                    id,
-                    username,
-                    name,
-                    biography,
-                    profile_picture_url,
-                    followers_count,
-                    follows_count,
-                    media_count
-                }',
-        ])->getContent();
+        $user = new User($this->user_id);
+        $accounts = $user->getAccounts([
+            'id',
+            'name',
+            'access_token',
+            'category',
+            'followers_count',
+            'verification_status',
+            'instagram_business_account{id,username,name,biography,profile_picture_url,followers_count,follows_count,media_count}',
 
-        return collect($accounts['data'])
-            ->filter(fn ($account) => isset($account['instagram_business_account']['id']))
-            ->map(function ($account) {
+        ]);
+
+        return collect($accounts->getArrayCopy())
+            ->filter(fn(Page $page) => isset($page->getData()['instagram_business_account']['id']))
+            ->map(function (Page $page) {
+                $account = $page->getData();
                 $ig = $account['instagram_business_account'];
 
                 return [
@@ -64,5 +60,29 @@ class Client
                     'access_token' => $account['access_token'],
                 ];
             });
+    }
+
+    public function getAllMedia(): array
+    {
+        $igUser = new IGUser($this->user_id);
+        $cursor = $igUser->getMedia([
+            'id',
+            'caption',
+            'media_type',
+            'media_product_type',
+            'media_url',
+            'thumbnail_url',
+            'permalink',
+            'timestamp',
+            'like_count',
+            'comments_count',
+        ]);
+        $cursor->setUseImplicitFetch(true);
+        $allMedia = [];
+        foreach ($cursor as $media) {
+            $allMedia[] = $media->exportAllData();
+        }
+
+        return $allMedia;
     }
 }